requirements.txt
#bunnet==1.3.0
fastapi==0.111.0
openai==0.27.10
pydantic==2.7.1
PyJWT==2.8.0
pymongo==4.6.3
Requests==2.32.2
uvicorn==0.29.0
google-generativeai==0.5.3
